% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_impute.R
\docType{methods}
\name{aggregate_impute}
\alias{aggregate_impute}
\alias{aggregate_impute,ANY-method}
\alias{aggregate_impute,rawImputed-method}
\alias{aggregate_impute,aggregatedImputed-method}
\title{Aggregate an imputed dataset}
\usage{
aggregate_impute(object, grouping, fun, filter = list(), join, ...)

\S4method{aggregate_impute}{ANY}(object, grouping, fun, filter = list(), join, ...)

\S4method{aggregate_impute}{rawImputed}(object, grouping, fun, filter = list(), join, ...)

\S4method{aggregate_impute}{aggregatedImputed}(object, grouping, fun, filter = list(), join, ...)
}
\arguments{
\item{object}{A \code{rawImputed} object.}

\item{grouping}{A vector of variables names to group the aggregation on.}

\item{fun}{The function to aggregate.}

\item{filter}{An optional argument to filter the raw dataset before aggregation.
Will be passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}

\item{join}{An optional argument to filter the raw dataset based on a data.frame.
A \code{\link[dplyr:semi_join]{dplyr::semi_join()}} will be applied with \code{join} or each element
of \code{join} in case join is a list.}

\item{...}{Additional arguments.
See details.}
}
\description{
Aggregate an imputed dataset
}
\details{
You can pass a data.frame to \code{minimum}.
This data.frame should contain the same grouping columns as \code{grouping} and
one additional column with the minimum value for each group.
The aggregated values will be clamped to the minimum value for each group.
}
\examples{
dataset <- generate_data(n_year = 10, n_site = 50, n_run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
imputed <- impute(data = dataset, model = model)
aggregate_impute(imputed, grouping = c("Year", "Period"), fun = sum)
}
