% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregatedImputation_class.R
\docType{class}
\name{aggregatedImputed-class}
\alias{aggregatedImputed-class}
\title{The aggregatedimputed class
Holds an aggregated imputation data set}
\description{
The aggregatedimputed class
Holds an aggregated imputation data set
}
\section{Slots}{

  \describe{
   \item{\code{Covariate}}{A data.frame with the covariates}
   \item{\code{Imputation}}{a matrix wih aggregated imputed values}
  }
}

