% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_imputation.R
\name{average_imputation}
\alias{average_imputation}
\title{Calculate the multiple imputation average over some imputation sets}
\usage{
average_imputation(data, model.fun, rhs, model.args, extractor, extractor.args)
}
\arguments{
\item{data}{a \code{data.frame} with the imputations and the covariates}

\item{model.fun}{the function to apply on each imputation set}

\item{rhs}{the right hand side of the model}

\item{model.args}{an optional list of arguments to pass to the model function}

\item{extractor}{a function which return a \code{matrix} or \code{data.frame}. The first column should contain the estimate, the second the standard error of the estimate}

\item{extractor.args}{an optional list of arguments to pass to the extractor function}
}
\description{
Calculate the multiple imputation average over some imputation sets
}

