% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{generateData}
\alias{generateData}
\alias{missingAtRandom}
\alias{missingCurrentCount}
\alias{missingObserved}
\alias{missingVolunteer}
\title{Deprecated functions}
\usage{
generateData(
  intercept = 2,
  n.year = 24,
  n.period = 6,
  n.site = 20,
  year.factor = FALSE,
  period.factor = FALSE,
  site.factor = FALSE,
  trend = 0.01,
  sd.rw.year = 0.1,
  amplitude.period = 1,
  mean.phase.period = 0,
  sd.phase.period = 0.2,
  sd.site = 1,
  sd.rw.site = 0.02,
  sd.noise = 0.01,
  size = 2,
  n.run = 10,
  as.list = FALSE,
  details = FALSE
)

missingAtRandom(
  dataset,
  proportion = 0.25,
  count.variable = "Count",
  observed.variable = "Observed"
)

missingCurrentCount(
  dataset,
  proportion = 0.25,
  count.variable = "Count",
  observed.variable = "Observed"
)

missingObserved(
  dataset,
  count.variable = "Count",
  observed.variable = "Observed",
  site.variable = "Site",
  year.variable = "Year",
  period.variable = "Period"
)

missingVolunteer(
  dataset,
  proportion = 0.25,
  count.variable = "Count",
  observed.variable = "Observed",
  year.variable = "Year",
  site.variable = "Site",
  max.count = 100
)
}
\arguments{
\item{intercept}{The global mean on the log-scale.}

\item{n.year}{The number of years.}

\item{n.period}{The number of periods.}

\item{n.site}{The number of sites.}

\item{year.factor}{Convert year to a factor.
Defaults to \code{FALSE}.}

\item{period.factor}{Convert period to a factor.
Defaults to \code{FALSE}.}

\item{site.factor}{Convert site to a factor.
Defaults to \code{FALSE}.}

\item{trend}{The long-term linear trend on the log-scale.}

\item{sd.rw.year}{The standard deviation of the year effects on the
log-scale.}

\item{amplitude.period}{The amplitude of the periodic effect on the
log-scale.}

\item{mean.phase.period}{The mean of the phase of the periodic effect among
years.
Defaults to \code{0}.}

\item{sd.phase.period}{The standard deviation of the phase of the periodic
effect among years.}

\item{sd.site}{The standard deviation of the site effects on the log-scale.}

\item{sd.rw.site}{The standard deviation of the random walk along year per
site on the log-scale.}

\item{sd.noise}{The standard deviation of the noise effects on the log-scale.}

\item{size}{The size parameter of the negative binomial distribution.}

\item{n.run}{The number of runs with the same mu.}

\item{as.list}{Return the dataset as a list rather than a data.frame.
Defaults to \code{FALSE}.}

\item{details}{Add variables containing the year, period and site effects.
Defaults tot \code{FALSE}.}

\item{dataset}{A dataset to a the observation with missing data.}

\item{proportion}{The proportion of observations that will be missing.}

\item{count.variable}{The name of the variable holding the counts.}

\item{observed.variable}{The name of the variable holding the observed values
= either count or missing.}

\item{site.variable}{The name of the variable holding the sites.}

\item{year.variable}{The name of the variable holding the years.}

\item{period.variable}{The name of the variable holding the period.}

\item{max.count}{The maximum count.}
}
\description{
These functions will be removed from \code{multimput} in the future.
}
