% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example}
\alias{example}
\title{The indices for the example dataset}
\format{A data frame with 240 rows and 7 variables}
\usage{
data(example)
}
\description{
Data for fig 9 in Onkelinx et al
}
\details{
\itemize{
  \item \code{Estimate} The point estimate of the index
  \item \code{LCL} The lower confidence limit of the index
  \item \code{UCL} The upper confidence limit of the index
  \item \code{Year} The year of the index
  \item \code{Method} A factor indicating the imputation technique
  \item \code{Type} A factor indicating if the index is based on the observed, augmented or complete data
  \item \code{Family} A factor indicating if the index is based on the negative binomial or Poisson distribution
}
}
\keyword{datasets}

