% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\docType{methods}
\name{extractor}
\alias{extractor}
\alias{extractor,ANY-method}
\title{Extract parameter estimated from a model}
\usage{
extractor(model, ...)

\S4method{extractor}{ANY}(model, ...)
}
\arguments{
\item{model}{model to extract the estimate and standard error of parameters}

\item{...}{other arguments}
}
\description{
Extract parameter estimated from a model
}

