% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate simulated data}
\usage{
generateData(intercept = 2, n.year = 24, n.period = 6, n.site = 20,
  year.factor = FALSE, period.factor = FALSE, site.factor = FALSE,
  trend = 0.01, sd.rw.year = 0.1, amplitude.period = 1,
  mean.phase.period = 0, sd.phase.period = 0.2, sd.site = 1,
  sd.rw.site = 0.02, sd.noise = 0.01, size = 2, n.run = 10,
  as.list = FALSE, details = FALSE)
}
\arguments{
\item{intercept}{the global mean on the log-scale}

\item{n.year}{the number of years}

\item{n.period}{the number of periods}

\item{n.site}{the number of sites}

\item{year.factor}{convert year to a factor. Defaults to FALSE}

\item{period.factor}{convert period to a factor. Defaults to FALSE}

\item{site.factor}{convert site to a factor. Defaults to FALSE}

\item{trend}{the longterm linear trend on the log-scale}

\item{sd.rw.year}{the standard deviation of the year effects on the log-scale}

\item{amplitude.period}{the amplitude of the periodic effect on the log-scale}

\item{mean.phase.period}{the mean of the phase of the periodic effect among years. Defaults to 0.}

\item{sd.phase.period}{the standard deviation of the phase of the periodic effect among years}

\item{sd.site}{the standard deviation of the site effects on the log-scale}

\item{sd.rw.site}{the standard deviation of the random walk along year per site on the log-scale}

\item{sd.noise}{the standard deviation of the noise effects on the log-scale}

\item{size}{the size parameter of the negative binomial distribution}

\item{n.run}{the number of runs with the same mu}

\item{as.list}{return the dataset as a list rather than a data.frame. Defaults to FALSE}

\item{details}{add variables containing the year, period and site effects. Defaults tot FALSE}
}
\value{
A \code{data.frame} with five variables. \code{Year}, \code{Month} and \code{Site} are factors identifying the location and time of monitoring. \code{Mu} is the true mean of the negative binomial distribution in the original scale. \code{Count} are the simulated counts.
}
\description{
Generate data for a regural monitoring design. The counts follow a negative binomial distribution with given size paramters and the true mean mu depending on a year, period and site effect. All effects are independent fro each other and have, on the log-scale, a normal distribution with zero mean and given standard deviation.
}

