% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdle_impute.R
\name{hurdle_impute}
\alias{hurdle_impute}
\title{Combine two models into a hurdle model}
\usage{
hurdle_impute(presence, count)
}
\arguments{
\item{presence}{the \code{rawImputed} object for the presence.}

\item{count}{the \code{rawImputed} object for counts.}
}
\description{
Multiplies the imputed values for the \code{presence} model with those of the
\code{count} model.
Please make sure that the order of the observations in both models is
identical.
The resulting object will contain the union of the covariates of both models.
Variables with the same name and different values get a \code{presence_} or
\code{count_} prefix.
}
