% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputedTotals.R
\name{imputedTotals}
\alias{imputedTotals}
\title{Calculate the totals over an imputed dataset}
\usage{
imputedTotals(data, imputations, variable, rhs)
}
\arguments{
\item{data}{the orginal dataset with missing values}

\item{imputations}{the multiple imputations generated by e.g. \code{\link{imputeINLA}}}

\item{variable}{the name of the variable for which the imputations are calculated}

\item{rhs}{the right hand side of the formula for the aggregation}
}
\value{
a dataset containing the totals for each imputations. The first few columns will contains the \code{rhs} variables.
}
\description{
Calculate the totals over an imputed dataset
}

