% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeINLA.R
\name{imputeINLA}
\alias{imputeINLA}
\title{impute missing data using multiple imputation}
\usage{
imputeINLA(data, formula = Observed ~ Year + Month + f(Site, model = "iid"),
  n.sim = 499, family = c("nbinomial", "poisson"))
}
\arguments{
\item{data}{a \code{data.frame} containing the variables \code{Year}, \code{Month}, \code{Site} and \code{Observed}. The missing values of \code{Observed} are imputed by the algorithm.}

\item{formula}{A formula defining the model to use for the imputation}

\item{n.sim}{the number of simulations}

\item{family}{"nbinomial" for a negative binomial distribution or "poisson" for a Poisson distribution}
}
\value{
A matrix with one row for each missing value. Each column is on imputation.
}
\description{
impute missing data using multiple imputation
}
\note{
Deprecated function. Only relevant to run the code in the folder \code{hpc} and \code{paper_simulation}.
}

