% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeTruth.R
\name{imputeTruth}
\alias{imputeTruth}
\title{impute missing data using multiple imputation}
\usage{
imputeTruth(data, size, mean = "Mu", respons = "Observed", n.sim = 499)
}
\arguments{
\item{data}{a \code{data.frame} containing the variables \code{Year}, \code{Month}, \code{Site} and \code{Observed}. The missing values of \code{Observed} are imputed by the algorithm.}

\item{size}{the size of the negative binomial distribution}

\item{mean}{the name of the variable in \code{data}, holding the true mean}

\item{respons}{the name of the variable in \code{data}, holding the observed values}

\item{n.sim}{the number of simulations}
}
\value{
A matrix with one row for each missing value. Each column is on imputation.
}
\description{
impute missing data using multiple imputation
}
\note{
Deprecated function. Only relevant to run the code in the folder \code{hpc} and \code{paper_simulation}.
}

