% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeUnderhill.R
\name{imputeUnderhill}
\alias{imputeUnderhill}
\title{impute missing data}
\usage{
imputeUnderhill(data, formula = Observed ~ Year + Month + Site, initial = 0,
  family = c("nbinomial", "poisson"))
}
\arguments{
\item{data}{a \code{data.frame} containing the variables \code{Year}, \code{Month}, \code{Site} and \code{Observed}. The missing values of \code{Observed} are imputed by the algorithm.}

\item{formula}{A formula defining the model to use for the imputation}

\item{initial}{the initial value by which the missing values are replaced}

\item{family}{"nbinomial" for a negative binomial distribution or "poisson" for a Poisson distribution}
}
\value{
A list with two elements: \code{data} with imputed values and \code{iterations} which is the number of iterations
}
\description{
impute missing data
}

