% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_at_random.R
\name{missing_at_random}
\alias{missing_at_random}
\title{Generate missing data at random}
\usage{
missing_at_random(
  dataset,
  proportion = 0.25,
  count_variable = "Count",
  observed_variable = "Observed"
)
}
\arguments{
\item{dataset}{A dataset to a the observation with missing data.}

\item{proportion}{The proportion of observations that will be missing.}

\item{count_variable}{The name of the variable holding the counts.}

\item{observed_variable}{The name of the variable holding the observed values
= either count or missing.}
}
\description{
The observed values will be either equal to the counts or missing.
The probability of missing is the inverse of the counts + 1.
}
