% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_observed.R
\name{missing_observed}
\alias{missing_observed}
\title{Generate missing data based on the observed patterns in the real dataset.}
\usage{
missing_observed(
  dataset,
  count_variable = "Count",
  observed_variable = "Observed",
  site_variable = "Site",
  year_variable = "Year",
  period_variable = "Period"
)
}
\arguments{
\item{dataset}{A dataset to a the observation with missing data.}

\item{count_variable}{The name of the variable holding the counts.}

\item{observed_variable}{The name of the variable holding the observed values
= either count or missing.}

\item{site_variable}{The name of the variable holding the sites.}

\item{year_variable}{The name of the variable holding the years.}

\item{period_variable}{The name of the variable holding the period.}
}
\description{
The observed values will be either equal to the counts or missing.
The probability of missing is the inverse of the counts + 1.
}
