% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_volunteer.R
\name{missing_volunteer}
\alias{missing_volunteer}
\title{Generate missing data mimicking choices made by volunteers.}
\usage{
missing_volunteer(
  dataset,
  proportion = 0.25,
  count_variable = "Count",
  observed_variable = "Observed",
  year_variable = "Year",
  site_variable = "Site",
  max_count = 100
)
}
\arguments{
\item{dataset}{A dataset to a the observation with missing data.}

\item{proportion}{The proportion of observations that will be missing.}

\item{count_variable}{The name of the variable holding the counts.}

\item{observed_variable}{The name of the variable holding the observed values
= either count or missing.}

\item{year_variable}{The name of the variable holding the years.}

\item{site_variable}{The name of the variable holding the sites.}

\item{max_count}{The maximum count.}
}
\description{
The observed values will be either equal to the counts or missing.
The probability of missing is the inverse of the counts + 1.
}
