% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_impute.R
\docType{methods}
\name{model_impute}
\alias{model_impute}
\alias{model_impute,ANY-method}
\alias{model_impute,aggregatedImputed-method}
\title{Model an imputed dataset}
\usage{
model_impute(
  object,
  model.fun,
  rhs,
  model.args,
  extractor,
  extractor.args,
  filter,
  mutate
)

\S4method{model_impute}{ANY}(
  object,
  model.fun,
  rhs,
  model.args,
  extractor,
  extractor.args,
  filter,
  mutate
)

\S4method{model_impute}{aggregatedImputed}(
  object,
  model.fun,
  rhs,
  model.args,
  extractor,
  extractor.args,
  filter,
  mutate
)
}
\arguments{
\item{object}{The imputed dataset.}

\item{model.fun}{The function to apply on each imputation set.}

\item{rhs}{The right hand side of the model.}

\item{model.args}{An optional list of arguments to pass to the model
function.}

\item{extractor}{A function which return a \code{matrix} or \code{data.frame}.
The first column should contain the estimate,
the second the standard error of the estimate.}

\item{extractor.args}{An optional list of arguments to pass to the \code{extractor} function.}

\item{filter}{An optional argument to filter the raw dataset before aggregation.
Will be passed to \code{\link[dplyr]{filter}}.}

\item{mutate}{An optional argument to alter the aggregated dataset.
Will be passed to the \code{.dots} argument of \code{\link[dplyr]{mutate}}.
This is mainly useful for simple convertions, e.g. factors to numbers and
viceversa.}
}
\description{
Model an imputed dataset
}
\examples{
dataset <- generateData(n.year = 10, n.site = 50, n.run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
imputed <- impute(data = dataset, model = model)
aggr <- aggregate_impute(imputed, grouping = c("Year", "Period"), fun = sum)
extractor <- function(model) {
  summary(model)$coefficients[, c("Estimate", "Std. Error")]
}
model_impute(
  object = aggr,
  model.fun = lm,
  rhs = "0 + factor(Year)",
  extractor = extractor
)
}
