% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_imputed_class.R
\docType{class}
\name{rawImputed-class}
\alias{rawImputed-class}
\title{The \code{rawImputed} class
Holds a dataset and imputed values}
\description{
The \code{rawImputed} class
Holds a dataset and imputed values
}
\section{Slots}{

\describe{
\item{\code{Data}}{A data.frame with the data.}

\item{\code{Response}}{A character holding the name of the response variable.}

\item{\code{Imputation}}{A matrix with imputed values.}

\item{\code{Extra}}{A data.frame with extra data to add to the imputations.
This data is not used in the imputation model.
It must contain the same variables as the original data.}
}}

