% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{results.inla}
\alias{results.inla}
\title{The Monte Carlo simualtion results using INLA}
\format{A data frame with 87600 rows and 12 variables}
\usage{
data(results.inla)
}
\description{
Data for fig 7 and 8 in Onkelinx et al
}
\details{
\itemize{
  \item \code{Estimate} The point estimate of the \eqn{API} index
  \item \code{Std. Error} The standard error of the \eqn{API} index
  \item \code{Year} The year of the index
  \item \code{2.5 \%} The lower confidence limit of the \eqn{API} index
  \item \code{97.5 \%} The upper confidence limit of the \eqn{API} index
  \item \code{Run} The id of the simulated dataset
  \item \code{Pattern} Pattern of missingness: 0 = missing completely at random (MCAR), 1 = observed
  \item \code{Missing} The proportions of missingness: 0 = observed, 1 = 1\%, 2 = 5\%, 3 = 25\%, 4 = 50\%, 5 = 75\%
  \item \code{Model} The imputation model: 0 = basic, 1 = true mean, 3 = complex
  \item \code{Imputations} The number ofimputations
}
}
\keyword{datasets}

