% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{results.trim}
\alias{results.trim}
\title{The Monte Carlo simualtion results using birdSTATs and TRIM}
\format{A data frame with 28800 rows and 10 variables}
\usage{
data(results.trim)
}
\description{
Data for fig 4, 5 and 6 in Onkelinx et al
}
\details{
\itemize{
  \item \code{Year} The year of the index
  \item \code{Estimate} The point estimate of the $APA$ index
  \item \code{Std. Error} The standard error of the $APA$ index
  \item \code{2.5 \%} The lower confidence limit of the $APA$ index
  \item \code{97.5 \%} The upper confidence limit of the $APA$ index
  \item \code{Run} The id of the simulated dataset
}
}
\keyword{datasets}

