% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{results.truth}
\alias{results.truth}
\title{The Monte Carlo simualtion results using the complete data}
\format{A data frame with 25200 rows and 12 variables}
\usage{
data(results.truth)
}
\description{
Data for fig 3 to 8 in Onkelinx et al
}
\details{
\itemize{
  \item \code{Estimate} The point estimate of the \eqn{API} index
  \item \code{Std. Error} The standard error of the \eqn{API} index
  \item \code{2.5 \%} The lower confidence limit of the \eqn{API} index
  \item \code{97.5 \%} The upper confidence limit of the \eqn{API} index
  \item \code{Estimatebis} The point estimate of the \eqn{APA} index
  \item \code{Std. Errorbis} The standard error of the \eqn{APA} index
  \item \code{2.5 \%bis} The lower confidence limit of the \eqn{APA} index
  \item \code{97.5 \%bis} The upper confidence limit of the \eqn{APA} index
  \item \code{Year} The year of the index
  \item \code{Run} The id of the simulated dataset
  \item \code{Years} The number of years in the dataset: 0 = 24, 1 = 6, 2 = 12, 3 = 24
  \item \code{Sites} The number of sites in the dataset: 0 = 100, 1 = 25, 2 = 50, 3 = 100
}
}
\keyword{datasets}

