% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{results.underhill}
\alias{results.underhill}
\title{The Monte Carlo simualtion results using Underhill}
\format{A data frame with 19200 rows and 9 variables}
\usage{
data(results.underhill)
}
\description{
Data for fig 3 in Onkelinx et al
}
\details{
\itemize{
  \item \code{Estimate} The point estimate of the $API$ index
  \item \code{Std. Error} The standard error of the $API$ index
  \item \code{2.5 \%} The lower confidence limit of the $API$ index
  \item \code{97.5 \%} The upper confidence limit of the $API$ index
  \item \code{Year} The year of the index
  \item \code{Run} The id of the simulated dataset
  \item \code{Pattern} Pattern of missingness: 0 = missing completely at random (MCAR)
  \item \code{Missing} The proportions of missingness: 0 = observed
  \item \code{Method} The algorithm: UHM = original Underhill with mean as start, UH0, original Underhill with 0 as start, UAM = altered Underhill with mean as start, UA0 = altered Underhill with 0 as start
}
}
\keyword{datasets}

