% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeImputationGLMnb.R
\name{summarizeImputationGLM.nb}
\alias{summarizeImputationGLM.nb}
\title{summarize the imputed dataset with a glm model}
\usage{
summarizeImputationGLM.nb(data, rhs)
}
\arguments{
\item{data}{the multiple imputed dataset}

\item{rhs}{the right hand side of the glm model to summarize the imputed data}
}
\description{
summarize the imputed dataset with a glm model
}
\note{
Deprecated function. Only relevant to run the code in the folder \code{hpc} and \code{paper_simulation}.
}

