% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeImputationGLM.R
\name{summarizeImputationGLM}
\alias{summarizeImputationGLM}
\title{summarize the imputed dataset with a glm model}
\usage{
summarizeImputationGLM(data, rhs)
}
\arguments{
\item{data}{the multiple imputed dataset}

\item{rhs}{the right hand side of the glm model to summarize the imputed data}
}
\description{
summarize the imputed dataset with a glm model
}

