% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{waterfowl}
\alias{waterfowl}
\title{The observation pattern in the Flemish waterfowl dataset}
\format{A data frame with 77157 rows and 5 variables}
\usage{
data(waterfowl)
}
\description{
Data for fig 1 and 2 in Onkelinx et al
}
\details{
\itemize{
  \item \code{Site} Site ID
  \item \code{Winter} Winter ID
  \item \code{Period} ID of the month
  \item \code{Species} Number of observed species
  \item \code{Birds} Total number of birds
}
}
\keyword{datasets}

