% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregated_imputed_class.R
\docType{class}
\name{aggregatedImputed-class}
\alias{aggregatedImputed-class}
\title{The aggregatedimputed class
Holds an aggregated imputation data set}
\description{
The aggregatedimputed class
Holds an aggregated imputation data set
}
\section{Slots}{

\describe{
\item{\code{Covariate}}{A data.frame with the covariates.}
\item{\code{Imputation}}{A matrix wih aggregated imputed values.}
}
}

