% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example}
\alias{example}
\title{The indices for the example dataset}
\format{A data frame with 240 rows and 7 variables}
\usage{
data(example)
}
\description{
Data for fig 9 in Onkelinx et al
}
\keyword{datasets}

