% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate simulated data}
\usage{
generate_data(
  intercept = 2,
  n_year = 24,
  n_period = 6,
  n_site = 20,
  year_factor = FALSE,
  period_factor = FALSE,
  site_factor = FALSE,
  trend = 0.01,
  sd_rw_year = 0.1,
  amplitude_period = 1,
  mean_phase_period = 0,
  sd_phase_period = 0.2,
  sd_site = 1,
  sd_rw_site = 0.02,
  sd_noise = 0.01,
  size = 2,
  n_run = 10,
  as_list = FALSE,
  details = FALSE
)
}
\arguments{
\item{intercept}{The global mean on the log-scale.}

\item{n_year}{The number of years.}

\item{n_period}{The number of periods.}

\item{n_site}{The number of sites.}

\item{year_factor}{Convert year to a factor.
Defaults to \code{FALSE}.}

\item{period_factor}{Convert period to a factor.
Defaults to \code{FALSE}.}

\item{site_factor}{Convert site to a factor.
Defaults to \code{FALSE}.}

\item{trend}{The longterm linear trend on the log-scale.}

\item{sd_rw_year}{The standard deviation of the year effects on the
log-scale.}

\item{amplitude_period}{The amplitude of the periodic effect on the
log-scale.}

\item{mean_phase_period}{The mean of the phase of the periodic effect among
years.
Defaults to \code{0}.}

\item{sd_phase_period}{The standard deviation of the phase of the periodic
effect among years.}

\item{sd_site}{The standard deviation of the site effects on the log-scale.}

\item{sd_rw_site}{The standard deviation of the random walk along year per
site on the log-scale.}

\item{sd_noise}{The standard deviation of the noise effects on the log-scale.}

\item{size}{The size parameter of the negative binomial distribution.}

\item{n_run}{The number of runs with the same mu.}

\item{as_list}{Return the dataset as a list rather than a data.frame.
Defaults to \code{FALSE}.}

\item{details}{Add variables containing the year, period and site effects.
Defaults tot \code{FALSE}.}
}
\value{
A \code{data.frame} with five variables.
\code{Year}, \code{Month} and \code{Site} are factors identifying the location and time of
monitoring.
\code{Mu} is the true mean of the negative binomial distribution in the original
scale.
\code{Count} are the simulated counts.
}
\description{
Generate data for a regural monitoring design.
The counts follow a negative binomial distribution with given size parameters
and the true mean mu depending on a year, period and site effect.
All effects are independent from each other and have, on the log-scale, a
normal distribution with zero mean and given standard deviation.
}
