% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_generic.R, R/impute_lm.R
\docType{methods}
\name{impute}
\alias{impute}
\alias{impute,lm-method}
\title{Impute a dataset}
\usage{
impute(model, ..., n.imp = 19)

\S4method{impute}{lm}(model, data, ..., n.imp = 19)
}
\arguments{
\item{model}{model to impute the dataset}

\item{...}{other arguments. See details}

\item{n.imp}{the number of imputations. Defaults to 19}

\item{data}{The dataset holding both the observed and the missing values}
}
\description{
Impute a dataset
}

