% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_generic.R, R/impute_glmerMod.R,
%   R/impute_inla.R, R/impute_lm.R
\docType{methods}
\name{impute}
\alias{impute}
\alias{impute,ANY-method}
\alias{impute,glmerMod-method}
\alias{impute,inla-method}
\alias{impute,lm-method}
\title{Impute a dataset}
\usage{
impute(model, ..., n.imp = 19)

\S4method{impute}{ANY}(model, ..., n.imp = 19)

\S4method{impute}{glmerMod}(model, data, ..., n.imp)

\S4method{impute}{inla}(model, ..., n.imp = 19)

\S4method{impute}{lm}(model, data, ..., n.imp)
}
\arguments{
\item{model}{model to impute the dataset}

\item{...}{other arguments. See details}

\item{n.imp}{the number of imputations. Defaults to 19}

\item{data}{The dataset holding both the observed and the missing values}
}
\description{
Impute a dataset
}
\examples{
dataset <- generateData(n.year = 10, n.site = 50, n.run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
impute(model, dataset)
}
