% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_impute.R
\docType{methods}
\name{model_impute}
\alias{model_impute}
\alias{model_impute,ANY-method}
\alias{model_impute,aggregatedImputed-method}
\title{Model an imputed dataset}
\usage{
model_impute(
  object,
  model_fun,
  rhs,
  model_args,
  extractor,
  extractor_args,
  filter,
  mutate,
  ...
)

\S4method{model_impute}{ANY}(
  object,
  model_fun,
  rhs,
  model_args,
  extractor,
  extractor_args,
  filter,
  mutate,
  ...
)

\S4method{model_impute}{aggregatedImputed}(
  object,
  model_fun,
  rhs,
  model_args,
  extractor,
  extractor_args,
  filter,
  mutate,
  ...
)
}
\arguments{
\item{object}{The imputed dataset.}

\item{model_fun}{The function to apply on each imputation set.}

\item{rhs}{The right hand side of the model.}

\item{model_args}{An optional list of arguments to pass to the model
function.}

\item{extractor}{A function which return a \code{matrix} or \code{data.frame}.
The first column should contain the estimate,
the second the standard error of the estimate.}

\item{extractor_args}{An optional list of arguments to pass to the \code{extractor} function.}

\item{filter}{An optional argument to filter the raw dataset before aggregation.
Will be passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}

\item{mutate}{An optional argument to alter the aggregated dataset.
Will be passed to the \code{.dots} argument of\code{\link[dplyr:mutate]{dplyr::mutate()}}.
This is mainly useful for simple conversions, e.g. factors to numbers and
vice versa.}

\item{...}{currently ignored.}
}
\description{
Model an imputed dataset
}
\examples{
dataset <- generate_data(n_year = 10, n_site = 50, n_run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
imputed <- impute(data = dataset, model = model)
aggr <- aggregate_impute(imputed, grouping = c("Year", "Period"), fun = sum)
extractor <- function(model) {
  summary(model)$coefficients[, c("Estimate", "Std. Error")]
}
model_impute(
  object = aggr,
  model_fun = lm,
  rhs = "0 + factor(Year)",
  extractor = extractor
)
}
