% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_impute.R
\docType{methods}
\name{model_impute}
\alias{model_impute}
\alias{model_impute,ANY-method}
\alias{model_impute,aggregatedImputed-method}
\title{Model an imputed dataset}
\usage{
model_impute(object, model.fun, rhs, model.args, extractor.fun, extractor.args)

\S4method{model_impute}{ANY}(object, model.fun, rhs, model.args, extractor.fun,
  extractor.args)

\S4method{model_impute}{aggregatedImputed}(object, model.fun, rhs, model.args,
  extractor.fun, extractor.args)
}
\arguments{
\item{object}{the imputed dataset}

\item{model.fun}{the function to apply on each imputation set}

\item{rhs}{the right hand side of the model}

\item{model.args}{an optional list of arguments to pass to the model function.}

\item{extractor.fun}{a function which return a \code{matrix} or \code{data.frame}. The first column should contain the estimate, the second the standard error of the estimate. \code{\link{extractor}} will be used when missing.}

\item{extractor.args}{an optional list of arguments to pass to the extractor function.}
}
\description{
Model an imputed dataset
}
\examples{
dataset <- generateData(n.year = 10, n.site = 50, n.run = 1)
dataset$Count[sample(nrow(dataset), 50)] <- NA
model <- lm(Count ~ Year + factor(Period) + factor(Site), data = dataset)
imputed <- impute(data = dataset, model = model)
aggr <- aggregate_impute(imputed, grouping = c("Year", "Period"), fun = sum)
model_impute(object = aggr, model.fun = lm, rhs = "0 + factor(Year)")
}

