% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_imputed_class.R
\docType{class}
\name{rawImputed-class}
\alias{rawImputed-class}
\title{The rawimputed class
Holds a dataset and imputed values}
\description{
The rawimputed class
Holds a dataset and imputed values
}
\section{Slots}{

\describe{
\item{\code{Data}}{A data.frame with the data.}
\item{\code{Response}}{A character holding the name of the response variable.}
\item{\code{Minimum}}{An optional character holding the name of the variable
with the minimum.}
\item{\code{Imputation}}{A matrix wih imputed values.}
}
}

