% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawImputation_class.R
\docType{class}
\name{rawImputed-class}
\alias{rawImputed-class}
\title{The rawimputed class
Holds a dataset, an imputation model and imputed values}
\description{
The rawimputed class
Holds a dataset, an imputation model and imputed values
}
\section{Slots}{

  \describe{
   \item{\code{Model}}{The imputation model}
   \item{\code{Data}}{a data.frame with the data}
   \item{\code{Response}}{a character holding the name of the response variable}
   \item{\code{Imputation}}{a matrix wih imputed values}
  }
}

